requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.23.2
sqlalchemy>=2.0.23
pg8000>=1.30.1
psycopg2-binary>=2.9.9
pydantic[email]>=2.4.2
pydantic-settings>=2.0.3
python-multipart==0.0.9
httpx==0.27.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt==3.2.0
python-dotenv>=1.0.0
slowapi>=0.1.9
mercadopago>=2.2.0
sentry-sdk>=1.39.1
prometheus-fastapi-instrumentator>=6.1.0
boto3>=1.28.0

